Library items are now objects.
The Song object is derived from dict, all the others inherit from unicode.
The objects can be constructed from list items in the song dictionaries, in
that case the first item in the list is taken as the primary representation.
The full list is always accessible via the .all() method.